fix SegmentedRangeCache, modify test code

+++ lorc/test.sh
@@ -1,17 +1,35 @@
+#!/bin/bash
+
+# arg1 = run/profile
+mode=$1
+if [ -z "$mode" ]; then
+    echo "Usage: $0 <run|profile>"
+    exit 1
+fi
+if [ "$mode" != "run" ] && [ "$mode" != "profile" ]; then
+    echo "Invalid mode: $mode"
+    exit 1
+fi
+
 # compile 
-g++ -std=c++17 -g -O2 *.cc -o ./bin/testLORC 
+g++ -std=c++17 -g -O2 -fno-omit-frame-pointer -rdynamic *.cc -o ./bin/testLORC 
 
 # run
-# ./bin/testLORC
+if [ "$mode" == "run" ]; then
+    ./bin/testLORC
+fi
 
 # profile
-perf record -F 99 -g -o ./profile/profile.data ./bin/testLORC
+if [ "$mode" == "profile" ]; then
+    echo "Profiling..."
+    perf record -F 99 --call-graph dwarf -g -o ./profile/profile.data ./bin/testLORC
 
-# 生成火焰图（需提前安装FlameGraph工具）
-perf script -i ./profile/profile.data | \
+    # 生成火焰图（需提前安装FlameGraph工具）
+    perf script -i ./profile/profile.data | \
     stackcollapse-perf.pl | \
     flamegraph.pl > ./profile/lorc_flamegraph.svg
+    rm profile/profile.data
+fi
 
 # clean
-rm profile/profile.data
 rm bin/testLORC
